+++ Mana/TextureManager.h
@@ -20,8 +20,9 @@ public:
 
 	void ClearFromMap(std::string s_id);
 	bool Load(std::string fileName, std::string id, SDL_Renderer* pRenderer);
-	void Draw(std::string id, int x, int y, int p_width, int p_height, SDL_RendererFlip flip = SDL_FLIP_NONE);
-	void DrawRegion(std::string id, int p_srcX, int p_srcY, int p_Width, int p_Height, int p_destX, int p_destY, SDL_RendererFlip flip = SDL_FLIP_NONE);
+	void Draw(std::string id, float x, float y, int p_width, int p_height, SDL_RendererFlip flip = SDL_FLIP_NONE);
+	void DrawRegion(std::string id, float p_srcX, float p_srcY, int p_Width, int p_Height, float p_destX, float p_destY,
+		SDL_RendererFlip flip = SDL_FLIP_NONE);
 	
 	Vector2D GetTextureSize(std::string p_id);
